% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dict_idx_lookup}
\alias{dict_idx_lookup}
\title{Find the max indices given a list of edit distances.}
\usage{
dict_idx_lookup(
  list_of_edit_distances,
  words,
  dict_words,
  cutoff_levenstein,
  cutoff_LCS
)
}
\arguments{
\item{list_of_edit_distances}{A list of numeric vectors.}

\item{words}{CharacterVector. List of words.}

\item{dict_words}{CharacterVector. Word dictionary.}

\item{cutoff_levenstein}{Integer. Levenshtein Distance cutoff.}

\item{cutoff_LCS}{Double. Least Common Substring ratio cutoff.}
}
\value{
IntegerVector
}
\description{
Function is designed to find the min indices in a list of numeric vectors.
The numeric vectors consist of edit distances, numeric values.
}
