% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_words_Cpp}
\alias{get_words_Cpp}
\title{C++ wrapper around main fun in get_words.R}
\usage{
get_words_Cpp(str)
}
\arguments{
\item{str}{Character. For example, a sentence like "Mijn naam is Gerard. Ik ben een goochelaar".}
}
\value{
Character Vector.
}
\description{
Function extracts words from text.
}
