% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idx_lookup_min}
\alias{idx_lookup_min}
\title{Find the max indices given a list of edit distances.}
\usage{
idx_lookup_min(list_of_edit_distances)
}
\arguments{
\item{list_of_edit_distances}{A list of numeric vectors.}
}
\value{
Integer Vector
}
\description{
Function is designed to find the min indices in a list of numeric vectors.
The numeric vectors consist of edit distances, numeric values.
}
