% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{most_similar_levenstein}
\alias{most_similar_levenstein}
\title{Calculate for each string the normalized Levenshtein distance given a set of target strings.}
\usage{
most_similar_levenstein(strings, targets)
}
\arguments{
\item{strings}{String Vector.}

\item{targets}{String Vector.}
}
\description{
Calculate for each string the normalized Levenshtein distance given a set of target strings.
}
\examples{
words <- c("stomme", "goede")
targets <- c("goed", "stom", "ziek")
most_similar_levenstein(words, targets)
}
