% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{most_similar_word}
\alias{most_similar_word}
\title{Find the most similar word in the dictionary.}
\usage{
most_similar_word(words, dict_words, dict_scores, cutoff)
}
\arguments{
\item{words}{CharacterVector. Vector of words.}

\item{dict_words}{CharacterVector. A word dictionary.}

\item{dict_scores}{IntegerVector. Scores for each word in the dictionary.}

\item{cutoff}{Integer. The number of single-character edits required
to change a word in the target word.}
}
\description{
Find the most similar word in the dictionary.
}
