% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proximity_search_dict}
\alias{proximity_search_dict}
\title{Find the sentiment score of a word in the sentiment dictionary.}
\usage{
proximity_search_dict(words, dict_words, dict_scores, cutoff)
}
\arguments{
\item{words}{CharacterVector. Vector of words.}

\item{dict_words}{CharacterVector. A word dictionary.}

\item{dict_scores}{IntegerVector. Scores for each word in the dictionary.}

\item{cutoff}{Integer. The number of single-character edits required
to change a word in the target word.}
}
\value{
IntegerVector. Sentiment scores for each word, 0 if no match was found.
}
\description{
Find the sentiment score of a word in the sentiment dictionary.
}
\examples{
words <- c("stomme", "goede")
proximity_search_dict(words, dict$word, dict$score, cutoff = 2)
}
