% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_words}
\alias{remove_words}
\title{Remove words from a vector of words.}
\usage{
remove_words(words, word_list)
}
\arguments{
\item{words.}{A vector of single words.}

\item{word_list.}{A vector with words that you want to remove.}
}
\description{
Remove words from a vector of words.
}
