% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_dict}
\alias{search_dict}
\title{Load your dutch sentiment dictionary and search a word in it.}
\usage{
search_dict(w, words, scores)
}
\arguments{
\item{w}{CharacterVector. Words you want to test, each in a separate vector.}

\item{words}{CharacterVector. Words from your sentiment dictionary.}

\item{scores}{IntegerVector. Scores that belong to each word in your dict.}
}
\description{
Find the sentiment score of a word in the sentiment dictionary.
}
