% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_contains}
\alias{str_contains}
\title{Check whether words are in a word list.}
\usage{
str_contains(words, word_list)
}
\arguments{
\item{words}{Vector of Characters, each a word.}

\item{word_list}{Vector of Characters, each a word.}
}
\description{
Returns \code{TRUE} if a word or vector of words
are in a word list. Otherwise it returns FALSE.
}
